third_party/WebKit/LayoutTests/fast/table/layout-section-when-table-height-changes-3.html

<!DOCTYPE html>
<style>
table {
  height: 100%;
  width: 100%;
}
</style>
<p> crbug.com/691968: There should be no scrollbar. </p>
<div id="container" style="width: 400px; height: 400px; overflow: auto; overflow-x: hidden;">
  <table id="table" data-expected-height=200>
    <colgroup>
      <col style="width: 1161px;">
    </colgroup>
    <tbody>
      <tr>
        <td></td>
      </tr>
    </tbody>
  </table>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
  document.body.offsetTop;
  document.getElementById("container").style.height = "200px";
  checkLayout('#table');
</script>
